mcp/chrome/pyproject.toml
[project]
name = "chrome-mcp"
version = "0.1.0"
description = "Lightweight MCP Server for Chrome Browser Automation"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
urls = { Homepage = "https://github.com/tomdwipo/agent" }
keywords = ["chrome", "mcp", "browser", "automation", "selenium", "web"]
dependencies = [
    "mcp>=1.9.3",
    "selenium>=4.15.0",
    "undetected-chromedriver>=3.5.0",
    "webdriver-manager>=4.0.0",
    "pillow>=10.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "typing-extensions>=4.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]